initial full version (subject to testing)
